% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexcell.coords.R
\name{hexcell.coords}
\alias{hexcell.coords}
\title{Coordinates for a single hexagonal cell}
\usage{
hexcell.coords(cell.radius = 1, rotation = "flat-topped")
}
\arguments{
\item{cell.radius}{length from the ceter of the cell to the six corners}
}
\description{
Coordinates for a single hexagonal cell
}
\examples{
hexcell.coords(1)
}

