% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexgrid.plot.R
\name{hexgrid.plot}
\alias{hexgrid.plot}
\title{Plots hexagonal grids}
\usage{
hexgrid.plot(x.coords, y.coords, cell.coords = hexcell.coords(1), ...)
}
\arguments{
\item{x.coords}{matrix of x coordinates (columns) for each y value (rows)}

\item{y.coords}{matrix of y coordinates (rows) for each y value (columns)}

\item{cell.coords}{list of 2 (x and y, from hexcell.coords()) with length of 6, for each of the hexagonal corners}

\item{...}{passes arguments to polygon}
}
\description{
I will describe more late
}
\examples{
# Example 1, default where cell.radius = 1 and hexagon is flat-topped
   rad <- 1
   par(mfrow = c(2,2), mar = c(2,2,2,2), oma = rep(1,4), cex = 0.75)
   hd <- hexgrid.coords(5, offset ="even", cell.radius = rad, rotation = "flat-topped")
   hexgrid.plot(hd[,,1,1], hd[,,1,2], cell.coords = hexcell.coords(rad, rotation = "flat-topped"), col = rgb(0,0,1,.1), border = "grey50")
   axis(1, labels = T)
   axis(2, labels = T, las = 1)
   mtext("Flat-topped, even offset", side = 3, line = 0, cex = 0.75)
   text(0, 0, labels = "(0, 0)", cex = 0.75)
}

