% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexcell.coords.R
\name{hexcell.coords}
\alias{hexcell.coords}
\title{Coordinates for a single hexagonal cell}
\usage{
hexcell.coords(cell.radius = 1, rotation = "flat-topped")
}
\arguments{
\item{cell.radius}{length from the ceter of the cell to the six corners}

\item{rotation}{either of the two values: "flat-topped" or "point-topped"}
}
\value{
a list of with horizontal and vertical coordinates, respectively names x and y.
}
\description{
attempt
}
\examples{
# Example 1, default where cell.radius = 1 and hexagon is flat-topped
   hexcell.coords()

# Example 2, cell.radius = 2 and hexagon is point-topped
hexcell.coords(2, roataion = "point-topped")
}

